+++ frontend/src/Components/DataArea/VacationDetails/VacationDetails.tsx
@@ -22,6 +22,11 @@ function VacationDetails(): JSX.Element {
     },[]);
 
     async function deleteVacation(vacationID:number) {
+        // Display confirmation dialog
+        const isConfirmed = window.confirm("Are you sure you want to delete this vacation?");
+        // Don't delete if user clicks "cancel"
+        if (!isConfirmed) return;
+        
         try {
             await vacationService.deleteVacation(vacationID);
             notify.successMsg("Vacation successfully deleted!");
